+++ ecommerce-platform/basket-service/src/main/java/ecommerce/basket/core/Baskets.java
@@ -4,18 +4,19 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class Baskets {
 
-    private static final AtomicInteger idGenerator = new AtomicInteger(0);
+    private static final AtomicLong idGenerator = new AtomicLong(0);
 
-    private Map<Integer, Basket> baskets = new HashMap<>();
+    private Map<Long, Basket> baskets = new HashMap<>();
 
-    public Optional<Basket> get(Integer buyerId) {
+    public Optional<Basket> get(Long buyerId) {
         return Optional.ofNullable(baskets.get(buyerId));
     }
 
-    public Basket getOrCreate(Integer buyerId) {
+    public Basket getOrCreate(Long buyerId) {
         if (baskets.containsKey(buyerId) ) {
             return baskets.get(buyerId);
         } else {
@@ -23,14 +24,14 @@ public class Baskets {
         }
     }
 
-    public Basket create(Integer buyerId) {
+    public Basket create(Long buyerId) {
 
         Basket basket = new Basket(idGenerator.getAndAdd(1), buyerId);
         baskets.put(buyerId, basket);
         return basket;
     }
 
-    public Basket clear(Integer buyerId) {
+    public Basket clear(Long buyerId) {
 
         return baskets.remove(buyerId);
     }
