+++ ecommerce-platform/ecommerce-shared/src/main/java/ecommerce/shared/model/Product.java
@@ -2,18 +2,18 @@ package ecommerce.shared.model;
 
 public class Product {
 
-    private final Integer id;
+    private final Long id;
     private final String title;
     private final Double unitPrice;
 
-    public Product(Integer id, String title, Double unitPrice) {
+    public Product(Long id, String title, Double unitPrice) {
 
         this.id = id;
         this.title = title;
         this.unitPrice = unitPrice;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
