+++ ecommerce-platform/stock-service/src/main/java/ecommerce/stock/core/Stock.java
@@ -8,19 +8,19 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 public class Stock {
 
-    private Map<Integer, AtomicInteger> productCounts = new HashMap<>();
+    private Map<Long, AtomicInteger> productCounts = new HashMap<>();
 
-    public void add(Integer productId, Integer count) {
+    public void add(Long productId, Integer count) {
         AtomicInteger counts = productCounts.getOrDefault(productId, new AtomicInteger(0));
         counts.addAndGet(count);
         productCounts.putIfAbsent(productId, counts);
     }
 
-    public Integer getCount(Integer productId) {
+    public Integer getCount(Long productId) {
         return productCounts.getOrDefault(productId, new AtomicInteger(0)).get();
     }
 
-    public Integer checkout(Integer productId, Integer countToCheckout) {
+    public Integer checkout(Long productId, Integer countToCheckout) {
         AtomicInteger countInStock = productCounts.getOrDefault(productId, new AtomicInteger(0));
 
 //        if (countToCheckout > countInStock.get()) {
